utilise inheritance to reduce code duplication
